+++ src/novelystlib/views/__init__.py
@@ -5,6 +5,7 @@ basic_view -- provide a generic class for viewing tree element properties.
 chapter_view -- provide a class for viewing and editing chapter properties.
 character_view -- provide a class for viewing and editing character properties.
 no view -- Provide a class for clearing the right frame.
+normal_scene-view -- Provide a tkinter based class for viewing and editing "Normal" scene properties.
 notes_scene-view -- Provide a tkinter based class for viewing and editing "Notes" scene properties.
 project_view -- provide a tkinter based class for viewing and editing project properties.
 scene_view -- provide a tkinter based class for viewing and editing scene properties.