+++ src/novelystlib/views/todo_scene_view.py
@@ -1,18 +1,17 @@
 """Provide a class for viewing and editing "Todo" scene properties.
 
-Copyright (c) 2022 Peter Triesberger
+Copyright (c) 2023 Peter Triesberger
 For further information see https://github.com/peter88213/novelyst
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 import tkinter as tk
 from tkinter import ttk
 from pywriter.pywriter_globals import *
-from novelystlib.views.basic_view import BasicView
-from novelystlib.widgets.label_entry import LabelEntry
-from novelystlib.widgets.my_string_var import MyStringVar
+from novelystlib.views.scene_view import SceneView
+from novelystlib.widgets.folding_frame import FoldingFrame
 
 
-class TodoSceneView(BasicView):
+class TodoSceneView(SceneView):
     """Class for viewing and editing "Todo" scene properties.
           
     Public methods:
@@ -42,26 +41,23 @@ class TodoSceneView(BasicView):
         self._treeSelectBinding = None
         self._uiEscBinding = None
 
-        # 'Tags' entry.
-        self._tags = MyStringVar()
-        LabelEntry(self._elementInfoWindow, text=_('Tags'), textvariable=self._tags, lblWidth=self._LBL_X).pack(anchor=tk.W, pady=2)
-
-        ttk.Separator(self._elementInfoWindow, orient=tk.HORIZONTAL).pack(fill=tk.X)
+        ttk.Separator(self._viewWindow, orient=tk.HORIZONTAL).pack(fill=tk.X)
 
         #--- Frame for arc specific widgets.
-        self._plotFrame = ttk.Frame(self._elementInfoWindow)
-
-        # Associated scene title display.
-        self._sceneFrame = ttk.Frame(self._plotFrame)
-        self._sceneFrame.pack(anchor=tk.W, fill=tk.X)
-        self._associatedSceneTitle = ttk.Label(self._sceneFrame)
-        self._associatedSceneTitle.pack(anchor=tk.W, pady=2)
-        ttk.Button(self._sceneFrame, text=_('Assign scene'), command=self._pick_scene).pack(side=tk.LEFT, padx=1, pady=2)
-        ttk.Button(self._sceneFrame, text=_('Clear assignment'), command=self._clear_assignment).pack(side=tk.LEFT, padx=1, pady=2)
+        self._arcFrame = FoldingFrame(self._viewWindow, _('Arc'), self._toggle_arcFrame)
+        self._arcInnerFrame = ttk.Frame(self._arcFrame)
 
         # Arc display.
-        self._arc = ttk.Label(self._plotFrame)
-        self._arc.pack(anchor=tk.W, pady=2)
+        self._arc = ttk.Label(self._arcInnerFrame)
+        self._arc.pack(anchor=tk.W)
+
+        # Associated scene title display.
+        self._sceneFrame = ttk.Frame(self._arcInnerFrame)
+        self._sceneFrame.pack(anchor=tk.W, fill=tk.X)
+        self._associatedSceneTitle = tk.Label(self._sceneFrame, anchor=tk.W, bg='white')
+        self._associatedSceneTitle.pack(anchor=tk.W, pady=2, fill=tk.X)
+        ttk.Button(self._sceneFrame, text=_('Assign scene'), command=self._pick_scene).pack(side=tk.LEFT, padx=1, pady=2)
+        ttk.Button(self._sceneFrame, text=_('Clear assignment'), command=self._clear_assignment).pack(side=tk.LEFT, padx=1, pady=2)
 
     def set_data(self, element):
         """Update the widgets with element's data.
@@ -70,32 +66,30 @@ class TodoSceneView(BasicView):
         """
         super().set_data(element)
 
-        # 'Tags' entry.
-        if self._element.tags is not None:
-            self._tagsStr = list_to_string(self._element.tags)
-        else:
-            self._tagsStr = ''
-        self._tags.set(self._tagsStr)
-
         # Frame for arc point specific widgets.
+        if self._ui.kwargs['show_arcs']:
+            self._arcFrame.show()
+        else:
+            self._arcFrame.hide()
+
+        self._associatedScene = None
         if self._element.scnArcs:
             # Arc display.
-            self._arc.config(text=f'{_("Arc")}: {self._element.scnArcs}')
+            self._arc.config(text=self._element.scnArcs)
 
             # Associated scene display.
             try:
                 self._associatedScene = self._element.kwVar.get('Field_SceneAssoc', None)
                 sceneTitle = self._ui.novel.scenes[self._associatedScene].title
             except:
-                self._associatedScene = None
                 sceneTitle = ''
             self._associatedSceneTitle['text'] = sceneTitle
 
-            if not self._plotFrame.winfo_manager():
-                self._plotFrame.pack(pady=2, fill=tk.X)
+            if not self._arcInnerFrame.winfo_manager():
+                self._arcInnerFrame.pack(pady=2, fill=tk.X)
         else:
-            if self._plotFrame.winfo_manager():
-                self._plotFrame.pack_forget()
+            if self._arcInnerFrame.winfo_manager():
+                self._arcInnerFrame.pack_forget()
 
     def apply_changes(self):
         """Apply changes.
@@ -103,16 +97,9 @@ class TodoSceneView(BasicView):
         Extends the superclass method.
         """
         # 'Arc name' entry.
-        if  self._element.kwVar.get('Field_SceneAssoc', None) != self._associatedScene:
+        if self._element.kwVar.get('Field_SceneAssoc', None) != self._associatedScene:
             self._element.kwVar['Field_SceneAssoc'] = self._associatedScene
             self._ui.isModified = True
-
-        # 'Tags' entry.
-        newTags = self._tags.get()
-        if self._tagsStr or newTags:
-            if newTags != self._tagsStr:
-                self._element.tags = string_to_list(newTags)
-                self._ui.isModified = True
 
         super().apply_changes()
 
@@ -176,4 +163,13 @@ class TodoSceneView(BasicView):
         self.set_data(self._element)
         self._ui.tv.tree.see(self._lastSelected)
         self._ui.tv.tree.selection_set(self._lastSelected)
+
+    def _toggle_arcFrame(self, event=None):
+        """Hide/show the narrative arcs frame."""
+        if self._ui.kwargs['show_arcs']:
+            self._arcFrame.hide()
+            self._ui.kwargs['show_arcs'] = False
+        else:
+            self._arcFrame.show()
+            self._ui.kwargs['show_arcs'] = True